Add release profile to Cargo.toml, update Dockerfile, and introduce new Dockerfile.release
This commit enhances the build configuration by adding a release profile in `Cargo.toml` to enable link-time optimization and set panic behavior. The `Dockerfile` is updated to streamline the installation of the protobuf compiler, while a new `Dockerfile.release` is introduced for building the application in a multi-stage Docker setup. Additionally, a new `buf.yaml` file is added for managing Buf modules, and updates to GitHub workflows include a new `release.yml` for handling release processes and security scans.