chore: update dependencies and improve Kubernetes labels
This commit updates the `aws-sdk-s3` and `aws-smithy-checksums` dependencies to their latest versions in `Cargo.lock` and `Cargo.toml`. It also modifies Kubernetes labels in the `justfile` and Helm templates to include instance-specific identifiers, enhancing clarity in deployment configurations. Additionally, it removes unused SQLx query files and refines error messages in the `reader_service` and `pod_watcher` modules for better debugging.